rename BundleObservation classes
