+++ isis/src/control/objs/BundleAdjust/unitTest.cpp
@@ -50,7 +50,7 @@ Errors:
  void BundleAdjust::init(Progress *progress) {
     (image==false) for BundleImage* image = new BundleImage(camera, serialNumber, fileName);
           QString msg = "In BundleAdjust::init(): image " + fileName + "is null" + "\n";
-    (observation==false) for BundleObservation *observation = m_bundleObservations.addNew(image, observationNumber, instrumentId, m_bundleSettings);
+    (observation==false) for IsisBundleObservation *observation = m_bundleObservations.addNew(image, observationNumber, instrumentId, m_bundleSettings);
           QString msg = "In BundleAdjust::init(): observation " + observationNumber + "is null" + "\n";
