+++ isis/src/control/objs/BundleUtilities/BundleUtilities.truth
@@ -333,9 +333,9 @@ file name     =  "TestImageFileName"
 XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
 
 XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
-Testing BundleObservation...
-Constructing empty BundleObservation object...
-Constructing BundleObservation object from BundleImage...
+Testing IsisBundleObservation...
+Constructing empty IsisBundleObservation object...
+Constructing IsisBundleObservation object from BundleImage...
 Testing assignment operator to set this equal to itself...
 Testing assignment operator to create a new object...
 Testing copy constructor...
@@ -519,7 +519,7 @@ TWI  (dd)          0.00000000            0.00000000               0.00000000    
 0.0,0.0,0.0,N/A,N/A,0.0,0.0,0.0,N/A,N/A,0.0,0.0,0.0,N/A,N/A,0.0,0.0,0.0,FREE,0.00000000,0.0,0.0,0.0,FREE,0.00000000,0.0,0.0,0.0,FREE,0.00000000,
 init exterior orientiation successful?   "Yes"
 apply param corrections successful?
-**ERROR** Unable to apply parameter corrections to BundleObservation.
+**ERROR** Unable to apply parameter corrections to IsisBundleObservation.
 **ERROR** Instrument rotation is NULL, but pointing solve option is AnglesOnly.
 
     add another image...
@@ -528,9 +528,9 @@ TestImageFileName
 TestImage2FileName
 
 Testing exceptions...
-**ERROR** Unable to apply parameter corrections to BundleObservation.
+**ERROR** Unable to apply parameter corrections to IsisBundleObservation.
 **ERROR** Instrument position is NULL, but position solve option is PositionOnly.
-**ERROR** Unable to apply parameter corrections to BundleObservation.
+**ERROR** Unable to apply parameter corrections to IsisBundleObservation.
 **ERROR** Instrument rotation is NULL, but pointing solve option is AnglesOnly.
 XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
